server_backend/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.3.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
coverage==7.6.12
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Mailman==1.1.1
Flask-MySQLdb==2.0.0
Flask-Principal==0.4.0
Flask-Security==5.6.1
Flask-Security-Too==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
matplotlib==3.8.4
mirakuru==2.6.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.6
numpy==2.2.0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.3.6
plotly==5.22.0
pluggy==1.5.0
port-for==0.7.0
psutil==7.0.0
pycparser==2.22
PyMySQL==1.1.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-mysql==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.0
scipy==1.14.1
six==1.17.0
SQLAlchemy==2.0.39
SQLAlchemy-Utils==0.41.2
statsmodels==0.14.2
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
EOF < /dev/null
